refactor: remove redundant EventPublisherConfiguration and SpringEventPublisher
Remove the EventPublisherConfiguration and SpringEventPublisher classes
as they are no longer needed. The SpringEventPublisher is now imported
from s2.spring.core.publisher, which centralizes the event publishing
logic and reduces duplication. This change simplifies the configuration
and improves maintainability by using a shared core module for event
publishing.

test: update assertions in OrderBookDeciderImplTest

Replace `Assertions.assertThat` with `assertThat` for consistency and
readability in the test assertions. This change aligns the test code
with the preferred assertion style, making it more concise and easier
to read.

build: update dependencies to use s2-spring-core

Change the dependencies in build.gradle.kts files to use
s2-spring-core instead of s2-automate-core or other modules. This
ensures that the projects are using the correct core module, which
provides the necessary functionality and reduces dependency
complexity.
