+++ s2-spring/sourcing/s2-spring-boot-starter-sourcing/build.gradle.kts
@@ -5,7 +5,7 @@ plugins {
 }
 
 dependencies {
-	api(project(":s2-spring:storing:s2-spring-boot-starter-storing"))
+	api(project(":s2-spring:s2-spring-core"))
 
 	Dependencies.Spring.tx(::implementation)
 	Dependencies.Spring.autoConfigure(::implementation, ::kapt)
