+++ s2-spring/sourcing/s2-spring-boot-starter-sourcing/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,2 +1,1 @@
-s2.spring.automate.config.EventPublisherConfiguration
 s2.spring.automate.sourcing.config.RetryTaskChannelConfiguration
