--- s2-spring/storing/s2-spring-boot-starter-storing/src/main/kotlin/s2/spring/automate/persister/SpringEventPublisher.kt
@@ -1,15 +1,0 @@
-package s2.spring.automate.persister
-
-import org.springframework.context.ApplicationEventPublisher
-import org.springframework.context.PayloadApplicationEvent
-import org.springframework.stereotype.Component
-import s2.automate.core.appevent.publisher.AppEventPublisher
-
-@Component
-class SpringEventPublisher(
-	private val applicationEventPublisher: ApplicationEventPublisher,
-) : AppEventPublisher {
-	override fun <EVENT> publish(event: EVENT & Any) {
-		applicationEventPublisher.publishEvent(PayloadApplicationEvent(this, event))
-	}
-}
