--- s2-spring/storing/s2-spring-boot-starter-storing/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,0 @@
-s2.spring.automate.config.EventPublisherConfiguration
